Changed name of vogn and removed wagon
